+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Force `do` blocks that we know have to be `do` blocks to break. (Thanks to @yuki24 for the report.)
 
 ## [0.5.0] - 2019-02-13
 ### Added