+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.1] - 2019-02-13
 ### Changed
 - Force `do` blocks that we know have to be `do` blocks to break. (Thanks to @yuki24 for the report.)
 - Handle `command` and `command_call` nodes `do` blocks by forcing them to break. (Thanks to @kmcq for the report.)
@@ -125,7 +127,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - Initial release 🎉
 
-[Unreleased]: https://github.com/CultureHQ/add-to-calendar/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/CultureHQ/add-to-calendar/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/CultureHQ/add-to-calendar/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/CultureHQ/add-to-calendar/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/CultureHQ/add-to-calendar/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/CultureHQ/add-to-calendar/compare/v0.3.7...v0.4.0